+++ Microsoft.Toolkit.Uwp.UI.Controls.DataGrid/DataGrid/Automation/DataGridItemAutomationPeer.cs
@@ -160,7 +160,7 @@ namespace Microsoft.Toolkit.Uwp.UI.Automation.Peers
         /// <returns>A Point that represents the clickable space that is on the UIElement that is associated with this DataGridItemAutomationPeer.</returns>
         protected override Point GetClickablePointCore()
         {
-            return this.OwningRowPeer != null ? this.OwningRowPeer.GetClickablePoint() : new Point(double.NaN, double.NaN);
+            return this.OwningRowPeer != null ? this.OwningRowPeer.GetClickablePoint() : new Point(float.NaN, float.NaN);
         }
 
         /// <summary>
