+++ Microsoft.Toolkit.Uwp.UI.Controls/InfiniteCanvas/Controls/InfiniteCanvasVirtualDrawingSurface.Render.cs
@@ -46,7 +46,7 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
             */
         }
 
-        private Rect ScaleRect(Rect rect, double scale)
+        private Rect ScaleRect(Rect rect, float scale)
         {
             return new Rect(rect.X * scale, rect.Y * scale, rect.Width * scale, rect.Height * scale);
         }
@@ -107,7 +107,7 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
         private Rect GetDrawingBoundaries(Rect viewPort)
         {
             _visibleList.Clear();
-            double top = double.MaxValue, bottom = double.MinValue, left = double.MaxValue, right = double.MinValue;
+            float top = float.MaxValue, bottom = float.MinValue, left = float.MaxValue, right = float.MinValue;
 
             foreach (var drawable in _drawableList)
             {
@@ -136,12 +136,12 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
 
             if (toDraw.Height > Height)
             {
-                toDraw.Height = Height;
+                toDraw.Height = (float)Height;
             }
 
             if (toDraw.Width > Width)
             {
-                toDraw.Width = Width;
+                toDraw.Width = (float)Width;
             }
 
             return toDraw;
@@ -150,7 +150,7 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
         /*
         private Rect GetMaxDrawingsBoundaries()
         {
-            double top = double.MaxValue, bottom = double.MinValue, left = double.MaxValue, right = double.MinValue;
+            float top = float.MaxValue, bottom = float.MinValue, left = float.MaxValue, right = float.MinValue;
 
             foreach (var drawable in _drawableList)
             {
